remove hertz depenedency, as it has a huge bug; further refine run_timed and test it in cpu code

+++ src/cpu/instruction.rs
@@ -227,7 +227,7 @@ pub struct RealizedInstruction
     pub instruction: Instruction,
     pub opcode: u8,
     pub addr_mode: AddressingMode,
-    pub cycles: u8,
+    pub cycles: u16,
     pub bytes: u8,
 }
